Merge pull request #119 from vjpalodichuk/feature/add_aoc2017_day18
Feature/add aoc2017 day18

+++ aoc-2016/src/main/kotlin/com/capital7software/aoc/aoc2016aoc/days/Day12.kt
@@ -109,7 +109,7 @@ class Day12 : AdventOfCodeSolution {
   fun getValueInRegister(input: List<String>, register: String): Int {
     val computer = SmallComputer.buildSmallComputer(input)
     computer.run()
-    return computer[register]
+    return computer[register].toInt()
   }
 
   /**
@@ -130,8 +130,8 @@ class Day12 : AdventOfCodeSolution {
       destValue: Int
   ): Int {
     val computer = SmallComputer.buildSmallComputer(input)
-    computer[destRegister] = destValue
+    computer[destRegister] = destValue.toLong()
     computer.run()
-    return computer[register]
+    return computer[register].toInt()
   }
 }
